Backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-cors==3.0.10
flask-jwt-extended==4.3.1
Flask-PyMongo==2.3.0
requests==2.28.1
python-dotenv==1.0.0
dnspython==2.2.1
email-validator==1.2.1